Rename `Lewee.StateManagement` to `Lewee.Infrastructure.Fluxor`
