+++ sample-tests/Pizzeria.Store.StateManagement.Tests.Unit/MessageReceivedActionTests.cs
@@ -1,5 +1,5 @@
 using FluentAssertions;
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 using Pizzeria.Store.Contracts.Orders;
 using Pizzeria.Store.StateManagement.Orders.Actions;
 using Xunit;
