+++ sample/Pizzeria.Store.StateManagement/Orders/Actions/AddPizzaToOrderCompletedAction.cs
@@ -1,4 +1,4 @@
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 using Pizzeria.Store.Contracts.Orders;
 
 namespace Pizzeria.Store.StateManagement.Orders.Actions;
