+++ sample/Pizzeria.Store.StateManagement/Orders/Actions/StartOrderFailureAction.cs
@@ -1,4 +1,4 @@
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 
 namespace Pizzeria.Store.StateManagement.Orders.Actions;
