+++ sample/Pizzeria.Store.StateManagement/Pizzas/Actions/LoadPizzasFailureAction.cs
@@ -1,4 +1,4 @@
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 
 namespace Pizzeria.Store.StateManagement.Pizzas.Actions;
