+++ sample/Pizzeria.Store.StateManagement/Pizzas/Actions/LoadPizzasSuccessAction.cs
@@ -1,4 +1,4 @@
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 using Pizzeria.Store.Contracts.Pizzas;
 
 namespace Pizzeria.Store.StateManagement.Pizzas.Actions;
