+++ src/Lewee.Infrastructure.Fluxor/IQuerySuccessAction.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Query Success Action Interface
