+++ src/Lewee.Infrastructure.Fluxor/IRequestAction.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Query Action
