+++ src/Lewee.Infrastructure.Fluxor/IRequestErrorAction.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Request Error Action Interface
