+++ src/Lewee.Infrastructure.Fluxor/IRequestState.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Request State Interface
