+++ src/Lewee.Infrastructure.Fluxor/IRequestSuccessAction.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Request Success Action Interface
