+++ src/Lewee.Infrastructure.Fluxor/QueryEffects.cs
@@ -2,10 +2,10 @@
 using Correlate;
 using Fluxor;
 using Lewee.Common;
-using Lewee.StateManagement.Observability;
+using Lewee.Infrastructure.Fluxor.Observability;
 using Microsoft.Extensions.Logging;
 
-namespace Lewee.StateManagement;
+namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Request Effects
@@ -15,24 +15,22 @@ namespace Lewee.StateManagement;
 /// <typeparam name="TRequestAction">Request action type</typeparam>
 /// <typeparam name="TRequestSuccessAction">Request success action type</typeparam>
 /// <typeparam name="TRequestErrorAction">Request error action type</typeparam>
-/// <typeparam name="TMessageReceived">Request completed action type</typeparam>
-public abstract class CommandEffects<TState, TData, TRequestAction, TRequestSuccessAction, TRequestErrorAction, TMessageReceived>
+public abstract class QueryEffects<TState, TData, TRequestAction, TRequestSuccessAction, TRequestErrorAction>
     : RequestEffects<TState, TData, TRequestSuccessAction, TRequestErrorAction>
     where TState : RequestState<TData>, new()
     where TData : class
     where TRequestAction : IRequestAction, new()
-    where TRequestSuccessAction : IRequestSuccessAction, new()
+    where TRequestSuccessAction : IQuerySuccessAction<TData>, new()
     where TRequestErrorAction : IRequestErrorAction, new()
-    where TMessageReceived : IMessageReceivedAction<TData>, new()
 {
     /// <summary>
     /// Initializes a new instance of the
-    /// <see cref="CommandEffects{TState, TData, TRequestAction, TRequestSuccessAction, TRequestErrorAction, TMessageReceived}"/> class
+    /// <see cref="QueryEffects{TState, TData, TRequestAction, TRequestSuccessAction, TRequestErrorAction}"/> class
     /// </summary>
     /// <param name="state">State</param>
     /// <param name="correlationContextAccessor">Correlation context accessor</param>
     /// <param name="logger">Logger</param>
-    protected CommandEffects(
+    protected QueryEffects(
         IState<TState> state,
         ICorrelationContextAccessor correlationContextAccessor,
         ILogger logger)
@@ -47,7 +45,7 @@ public abstract class CommandEffects<TState, TData, TRequestAction, TRequestSucc
     /// <param name="dispatcher">Dispatcher</param>
     /// <returns>Asynchronous task</returns>
     [EffectMethod]
-    public virtual async Task OnCommandAsync(
+    public virtual async Task OnQueryAsync(
         [NotNull] TRequestAction action,
         [NotNull] IDispatcher dispatcher)
     {
@@ -57,10 +55,14 @@ public abstract class CommandEffects<TState, TData, TRequestAction, TRequestSucc
 
             try
             {
-                var result = await this.ExecuteCommandAsync(action, dispatcher);
+                var result = await this.ExecuteQueryAsync(action, dispatcher);
                 if (result.IsSuccess)
                 {
-                    dispatcher.Dispatch(new TRequestSuccessAction() { CorrelationId = action.CorrelationId });
+                    dispatcher.Dispatch(new TRequestSuccessAction
+                    {
+                        CorrelationId = action.CorrelationId,
+                        Data = result.Data!,
+                    });
 
                     return;
                 }
@@ -88,41 +90,12 @@ public abstract class CommandEffects<TState, TData, TRequestAction, TRequestSucc
     }
 
     /// <summary>
-    /// Command completed effect
-    /// </summary>
-    /// <param name="action">Action</param>
-    /// <param name="dispatcher">Dispatcher</param>
-    /// <returns>Asynchronous task</returns>
-    [EffectMethod]
-    public virtual async Task OnCommandCompletedAsync(
-        [NotNull] TMessageReceived action,
-        [NotNull] IDispatcher dispatcher)
-    {
-        using (this.Logger.BeginCorrelationIdScope(this.CorrelationContextAccessor, action))
-        {
-            await this.ExecuteCommandCompletedAsync(action, dispatcher);
-
-            this.Logger.LogDebug("Executing query request...completed");
-        }
-    }
-
-    /// <summary>
-    /// Executes the command
+    /// Executes the query
     /// </summary>
     /// <param name="action">Request action</param>
     /// <param name="dispatcher">Dispatcher</param>
     /// <returns>Asynchronous task containing a <see cref="Result"/> that represents the success or failure of the request</returns>
-    protected abstract Task<CommandResult> ExecuteCommandAsync(
+    protected abstract Task<QueryResult<TData>> ExecuteQueryAsync(
         [NotNull] TRequestAction action,
-        [NotNull] IDispatcher dispatcher);
-
-    /// <summary>
-    /// Executes the command completed
-    /// </summary>
-    /// <param name="action">Command completed action</param>
-    /// <param name="dispatcher">Dispatcher</param>
-    /// <returns>Asynchronous task</returns>
-    protected abstract Task ExecuteCommandCompletedAsync(
-        [NotNull] TMessageReceived action,
         [NotNull] IDispatcher dispatcher);
 }
