+++ src/Lewee.Infrastructure.Fluxor/RequestState.cs
@@ -1,4 +1,4 @@
-﻿namespace Lewee.StateManagement;
+﻿namespace Lewee.Infrastructure.Fluxor;
 
 /// <summary>
 /// Command State
