+++ tests/Lewee.Blazor.Tests.Integration/Infrastructure/MessageToActionMapper.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using Lewee.Blazor.Messaging;
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 using Lewee.Tests.Contracts;
 using Microsoft.Extensions.Logging;
