+++ tests/Lewee.Blazor.Tests.Integration/Infrastructure/OrderCreatedAction.cs
@@ -1,4 +1,4 @@
-using Lewee.StateManagement;
+using Lewee.Infrastructure.Fluxor;
 using Lewee.Tests.Contracts;
 
 namespace Lewee.Blazor.Tests.Integration.Infrastructure;
